backend/requirements.txt
asgiref==3.9.1
Django==5.0.6
django-filter==25.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
